+++ packages/util/src/index.ts
@@ -1,2 +1,4 @@
 export { parseUrl, getPublicId, getTransformations } from './lib/cloudinary';
+export type { ParseUrl } from './lib/cloudinary';
+
 export { encodeBase64, objectHasKey, sortByKey } from './lib/util';