+++ packages/util/src/lib/cloudinary.ts
@@ -1,21 +1,25 @@
 const REGEX_VERSION = /\/v\d+\//;
-const REGEX_URL = /https?:\/\/(?<host>[^\/]+)\/(?<cloudName>[^\/]+)\/(?<assetType>image|video|raw)\/(?<deliveryType>upload|fetch|private|authenticated|sprite|facebook|twitter|youtube|vimeo)\/?(?<signature>s\-\-[a-zA-Z0-9]+\-\-)?\/?(?<transformations>(?:[^_\/]+_[^,\/]+,?\/?)*\/)+(?<version>v\d+|\w{1,2})\/(?<id>[^\.^\s]+)(?<format>\.[a-zA-Z0-9]+$)?$/;
+const REGEX_URL = /https?:\/\/(?<host>[^\/]+)\/(?<cloudName>[^\/]+)\/(?<assetType>image|images|video|videos|raw|files)\/(?<deliveryType>upload|fetch|private|authenticated|sprite|facebook|twitter|youtube|vimeo)?\/?(?<signature>s\-\-[a-zA-Z0-9]+\-\-)?\/?(?<transformations>(?:[^_\/]+_[^,\/]+,?\/?)*\/)*(?<version>v\d+|\w{1,2})\/(?<publicId>[^\.^\s]+)(?<format>\.[a-zA-Z0-9]+$)?$/;
+const ASSET_TYPES_SEO = ['images', 'videos', 'files'];
+
 
 /**
  * parseUrl
  * @description
  */
 
-interface ParseUrl {
+export interface ParseUrl {
   assetType?: string;
   cloudName?: string;
   deliveryType?: string;
   format?: string;
   host?: string;
-  id?: string;
+  publicId?: string;
   signature?: string;
+  seoSuffix?: string;
   transformations?: Array<string>;
-  version?: string;
+  queryParams?: object;
+  version?: number;
 }
 
 export function parseUrl(src: string): ParseUrl | undefined {
@@ -29,11 +33,35 @@ export function parseUrl(src: string): ParseUrl | undefined {
     throw new Error(`Invalid src: Does not include version (Ex: /v1234/)`);
   }
 
-  const results = src.match(REGEX_URL);
+  const [baseUrl, queryString] = src.split('?');
 
-  const parts = {
+  const results = baseUrl.match(REGEX_URL);
+  const transformations = results?.groups?.transformations?.split('/').filter(t => !!t);
+
+  const parts: ParseUrl = {
     ...results?.groups,
-    transformations: results?.groups?.transformations.split('/').filter(t => !!t)
+    seoSuffix: undefined,
+    transformations: transformations || [],
+    queryParams: {},
+    version: results?.groups?.version ? parseInt(results.groups.version.replace('v', '')) : undefined
+  }
+
+  if ( queryString ) {
+    interface QueryParams {
+      [key: string]: string | undefined;
+    }
+
+    parts.queryParams = queryString.split('&').reduce((prev: QueryParams, curr: string) => {
+      const [key, value] = curr.split('=');
+      prev[key] = value;
+      return prev;
+    }, {});
+  }
+
+  if ( parts.assetType && ASSET_TYPES_SEO.includes(parts.assetType) ) {
+    const publicIdParts = parts.publicId?.split('/') || [];
+    parts.seoSuffix = publicIdParts.pop();
+    parts.publicId = publicIdParts.join('/');
   }
 
   return parts;
@@ -48,8 +76,8 @@ export function parseUrl(src: string): ParseUrl | undefined {
  */
 
 export function getPublicId(src: string): string | undefined {
-  const { id } = parseUrl(src) || {};
-  return id;
+  const { publicId } = parseUrl(src) || {};
+  return publicId;
 }